+++ Frameworks/Accounts.framework/ACDispatchCerberus.h
@@ -5,6 +5,6 @@
 @interface ACDispatchCerberus : NSObject {
 }
 
-+ (void)waitOnSemaphore:(id)arg1 withTimeout:(unsigned int)arg2;
++ (void)waitOnSemaphore:(id)arg1 withTimeout:(unsigned long long)arg2;
 
 @end
