+++ Frameworks/AddressBookUI.framework/ABBoundLayoutConstraint.h
@@ -12,7 +12,7 @@
 @property(copy) NSString * keyPath;
 @property(retain) id object;
 
-+ (id)constraintWithItem:(id)arg1 attribute:(int)arg2 relatedBy:(int)arg3 toItem:(id)arg4 attribute:(int)arg5 multiplier:(float)arg6 constantObject:(id)arg7 keyPath:(id)arg8;
++ (id)constraintWithItem:(id)arg1 attribute:(long long)arg2 relatedBy:(long long)arg3 toItem:(id)arg4 attribute:(long long)arg5 multiplier:(double)arg6 constantObject:(id)arg7 keyPath:(id)arg8;
 
 - (void)_boundValueDidChange;
 - (void)bindConstantToObject:(id)arg1 keyPath:(id)arg2;
