+++ Frameworks/AddressBookUI.framework/ABImageAdditions.h
@@ -6,8 +6,8 @@
 }
 
 + (void)forEachImageFormat:(id)arg1;
-+ (unsigned int)indexForImageFormat:(int)arg1;
-+ (id)newDictionaryFromRect:(struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })arg1;
-+ (struct CGRect { struct CGPoint { float x_1_1_1; float x_1_1_2; } x1; struct CGSize { float x_2_1_1; float x_2_1_2; } x2; })rectFromDictionary:(id)arg1;
++ (unsigned long long)indexForImageFormat:(int)arg1;
++ (id)newDictionaryFromRect:(struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })arg1;
++ (struct CGRect { struct CGPoint { double x_1_1_1; double x_1_1_2; } x1; struct CGSize { double x_2_1_1; double x_2_1_2; } x2; })rectFromDictionary:(id)arg1;
 
 @end
