+++ Frameworks/AddressBookUI.framework/ABPropertyFaceTimeAction.h
@@ -6,11 +6,11 @@
 
 @interface ABPropertyFaceTimeAction : ABPropertyAction {
     ABPropertyBestFaceTimeQuery *_bestFaceTimeQuery;
-    int _type;
+    long long _type;
 }
 
 @property(retain) ABPropertyBestFaceTimeQuery * bestFaceTimeQuery;
-@property int type;
+@property long long type;
 
 - (void)_queryFaceTimeStatus;
 - (id)bestFaceTimeQuery;
@@ -18,7 +18,7 @@
 - (id)initWithContact:(id)arg1 propertyItems:(id)arg2;
 - (void)performActionForItem:(id)arg1 sender:(id)arg2;
 - (void)setBestFaceTimeQuery:(id)arg1;
-- (void)setType:(int)arg1;
-- (int)type;
+- (void)setType:(long long)arg1;
+- (long long)type;
 
 @end
