+++ Frameworks/AddressBookUI.framework/ABPropertyPostalAddressEditingCell.h
@@ -17,8 +17,8 @@
 - (id)editorSeparatorColor;
 - (id)firstResponderItem;
 - (void)layoutChanged:(id)arg1;
-- (float)leftValueMargin;
-- (float)rightContentMargin;
+- (double)leftValueMargin;
+- (double)rightContentMargin;
 - (void)setAddressEditor:(id)arg1;
 - (void)setBackgroundColor:(id)arg1;
 - (void)setEditorSeparatorColor:(id)arg1;
