+++ Frameworks/CFNetwork.framework/NSURLAuthenticationChallengeInternal.h
@@ -7,13 +7,13 @@
 @interface NSURLAuthenticationChallengeInternal : NSObject {
     NSError *error;
     NSURLResponse *failureResponse;
-    int previousFailureCount;
+    long long previousFailureCount;
     NSURLCredential *proposedCredential;
     <NSURLAuthenticationChallengeSender> *sender;
     NSURLProtectionSpace *space;
 }
 
 - (void)dealloc;
-- (id)initWithProtectionSpace:(id)arg1 proposedCredential:(id)arg2 previousFailureCount:(int)arg3 failureResponse:(id)arg4 error:(id)arg5 sender:(id)arg6;
+- (id)initWithProtectionSpace:(id)arg1 proposedCredential:(id)arg2 previousFailureCount:(long long)arg3 failureResponse:(id)arg4 error:(id)arg5 sender:(id)arg6;
 
 @end
