+++ Frameworks/CloudKit.framework/CKModifyBadgeOperation.h
@@ -7,22 +7,22 @@
  */
 
 @interface CKModifyBadgeOperation : CKOperation {
-    unsigned int _badgeValue;
+    unsigned long long _badgeValue;
     id _modifyBadgeCompletionBlock;
 }
 
-@property unsigned int badgeValue;
+@property unsigned long long badgeValue;
 @property(copy) id modifyBadgeCompletionBlock;
 
 - (void).cxx_destruct;
 - (void)_finishOnCallbackQueueWithError:(id)arg1;
-- (unsigned int)badgeValue;
+- (unsigned long long)badgeValue;
 - (void)fillOutOperationInfo:(id)arg1;
 - (id)init;
-- (id)initWithBadgeValue:(unsigned int)arg1;
+- (id)initWithBadgeValue:(unsigned long long)arg1;
 - (id)modifyBadgeCompletionBlock;
 - (void)performCKOperation;
-- (void)setBadgeValue:(unsigned int)arg1;
+- (void)setBadgeValue:(unsigned long long)arg1;
 - (void)setModifyBadgeCompletionBlock:(id)arg1;
 
 @end
