+++ Frameworks/CoreFoundation.framework/CFPDCFDataBuffer.h
@@ -11,6 +11,6 @@
 - (void)dealloc;
 - (void)finalize;
 - (id)initWithCFData:(struct __CFData { }*)arg1;
-- (unsigned long)length;
+- (unsigned long long)length;
 
 @end
